NavBar
Se ha creado un menu de navegeación con las opciones ver historico, perfil, ajustes y cerrar sesión
